React Redux Timers
A sample HTML which can be used to create multiple times, clear or reset and delete them. All timers count down synchronously in the prototype. The document is based on React Redux Store and the UI and state are tied to the store. All actions are dispatched via the store.